==> notebooks/01-load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests
import os

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """
    Constructs the URL where the datafile is.

    Args:
    `year` (int) - year of URL
    `month` (int) - month of URL
    """
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

download_one_file_of_raw_data(year=2022, month=1)

Define raw and output directories.

data = f"{os.getcwd().split("notebooks")[0]}data/"
data_raw = f"{data}/raw/"
data_transformed = f"{data}/transformed/"

directories = [data_raw, data_transformed]

[os.makedirs(data) for data in directories if not os.path.exists(data)]
# if not os.path.exists(data_transformed):
#     os.makedirs(data_transformed)

Load the file into a pandas DataFrame.

import pandas as pd

rides = pd.read_parquet(f"{data_raw}rides_2022-01.parquet")

rides.head(20)


Keep necessary columns for the project.

rides = rides[["tpep_pickup_datetime", "PULocationID"]]

rides

Rename columns.

rides.rename(columns={
    "tpep_pickup_datetime" : "pickup_datetime", 
    "PULocationID": "pickup_location_id"
}, inplace=True)

rides

## Data validation ##

There is never too much data validation procedures. While there are no data engineers to help us, let's check the data quality. 

For instance, let's make sure that our data comes from the registered timestamp interval (January 2022).

rides["pickup_datetime"].describe()

There are rows that shouldn't be there. For instance, the earliest date is in 2008, while the latest is in May 2022. We should remove all this data.

# Remove rows whose pickup_datetime is earlier than Jan 2022 and after Feb 2022.
rides = rides[rides["pickup_datetime"] >= "2022-01-01"]
rides = rides[rides["pickup_datetime"] <= "2022-02-01"]
rides["pickup_datetime"].describe()

Let's store the data in a `transformed` subdirectory.

rides.to_parquet(f"{data_transformed}validated_rides_2022-01.parquet")
