==> notebooks/02-transform_raw_data_into_ts_data.ipynb <==
# 02 | Transform data into Time Series data #
In this notebook, we will import our validated raw data and transform it into a time series format.

import pandas as pd
import os

os.getcwd()

# Add input directory
data_in = f"{os.getcwd().split("notebooks")[0]}data/transformed/"
data_in

rides = pd.read_parquet(f"{data_in}validated_rides_2022-01.parquet")

rides.head(10)

Add a column that contains the rounded hour.

rides["pickup_hour"] = rides["pickup_datetime"].dt.floor("H")
rides

Group the data by pickup hour and location ID. We want to count how many rows (rides) per location ID and per pickup hour.

agg_rides = rides.groupby(["pickup_hour", "pickup_location_id"]).size().reset_index(name="rides") # Name new column as 'rides'
agg_rides

Add extra rows with `zero` if not rides were registered in a given day.

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """
    Add missing slots when no observations were registered.
    
    Args:
    `agg_rides` (`pd.DataFrame`) - Pandas DataFrame with time series data to be transformed.

    Returns:
    `output` (`pd.DataFrame`)
    """

    location_ids =  agg_rides["pickup_location_id"].unique() # Get unique ride ids
    full_range = pd.date_range(
        agg_rides["pickup_hour"].min(), agg_rides["pickup_hour"].max(), freq="h"
    ) # Get the min and max pickup_hour observations
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):

        # Keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides["pickup_location_id"] == location_id, ["pickup_hour", "rides"]]

        # Quick way to add missing datas with 0 in a Series
        agg_rides_i.set_index("pickup_hour", inplace=True) # Set 'pickup_hour' as index
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0) # Reindex using min and max datetimes as the full range, passing 0 to NaN values

        # Add back `location_id` columns
        agg_rides_i["pickup_location_id"] = location_id

        output = pd.concat([output, agg_rides_i])

    # Move the purchase_day from the index to a DataFrame column
    output = output.reset_index().rename(columns={"index": "pickup_hour"})

    return output



agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

## Visualize the data ##

from typing import Optional, List
import plotly.express as px

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
):
    """
    Plot a time series data

    Args:
    `rides` (`pd.DataFrame`) - Pandas DataFrame
    `locations` (`List`) - List of integers with locations
    """

    rides_to_plot = rides[rides["pickup_location_id"].isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color="pickup_location_id",
        template="none"
    )

    fig.show()



We will use `pickup_location_id` `43` (Central Park).

plot_rides(agg_rides_all_slots, locations=[43])

Let's store the data to use it in our next process.

agg_rides_all_slots.to_parquet(f"{data_in}ts_data_2022-01.parquet")
