==> notebooks/03-transform_ts_data_into_features_and_target.ipynb <==
# 03 | Transform Time Series into Features and Targets

import pandas as pd
import os
import numpy as np
from tqdm import tqdm

Import the time series data.

data_in = f"{os.getcwd().split("notebooks")[0]}data/transformed/"

ts_data = pd.read_parquet(f"{data_in}ts_data_2022-01.parquet")
ts_data

Implement transformation for `location_id` 43 (Central Park)

ts_data_one_location = ts_data.loc[ts_data["pickup_location_id"] == 43, :].reset_index(drop=True)
ts_data_one_location.head(15)

Create the `get_cutoff_indices` function to slide and slice our data using a moving 12h interval.

def get_cutoff_indices(data: pd.DataFrame, n_features: int, step_size: int) -> list:
    """
    Slices a time series Pandas DataFrame with a given number of features and step size.

    Args:
    `data` (`pd.DataFrame`) - time series Pandas DataFrame
    `n_features` - number of features used to slice the data
    `step_size` - size to slide the feature window. 

    Returns:
    `list` - List of tuples containing the cutoff indices (min, max, step)
    """

    stop_position = len(data) - 1

    # Start the first subsequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size

    return indices

n_features = 24
step_size = 1

indices = get_cutoff_indices(ts_data_one_location,
                             n_features,
                             step_size
                             )

indices[:5]

Implement the slicing using the indices.

n_examples = len(indices)
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape=(n_examples), dtype=np.float32)
pickup_hours = []

for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0] : idx[1]]["rides"].values
    y[i] = ts_data_one_location.iloc[idx[1] : idx[2]]["rides"].values[0]
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]["pickup_hour"])

We defined one `np.array` for the features (`x`), whose shape is the number of examples and the number of features. `y` will contain the targets, which is a one-dimensional vector with the number of observations. They are our targets.

To generate the features, and the targets, we will loop over the list of indices we have just generated. 
- For every index, we extract the first and the second values, which are the cutoff points where you extract the features.
- The mid one and the right side one are the cutoff index to extract the target

print(f"{x.shape = }")
print(f"{x=}")
print(f"{pickup_hours[:5]=}")

Transform a Numpy array to a pandas DataFrame.

features_one_location = pd.DataFrame(
    x,
    columns=[f"rides_previous_{i+1}_hour" for i in reversed(range(n_features))]
)

features_one_location

So now we have a DataFrame with our features. Now, we are going to apply a similar principle to our target, transforming it to a DataFrame too.

targets_one_location = pd.DataFrame(y,
                                    columns=[f"target_rides_next_hour"])
targets_one_location

Let's create a function that performs these transformation steps for **all locations**.

def transform_ts_data_into_features_and_target(
        ts_data: pd.DataFrame,
        input_seq_len: int,
        step_size: int
) -> pd.DataFrame:
    """
    Slices and transposes data from time series format into a (`features`,`target`)
    format that we can use to train supervised ML models.
    
    Args:
    `ts_data` (`pd.DataFrame`) - time series Pandas DataFrame
    `input_seq_len` (`int`) - length of the slice to be performed
    `step_size` - size to slide the feature window. 

    Returns:
    `pd.DataFrame` for features and `pd.DataFrame` for targets
    """    
    assert set(ts_data.columns) == {"pickup_hour", "rides", "pickup_location_id"}  # Validates variables

    location_ids = ts_data["pickup_location_id"].unique() # Store unique ids
    features = pd.DataFrame() # Create empty DataFrame for features
    targets = pd.DataFrame() # Create empty DataFrame for features

    for location_id in tqdm(location_ids):

        # Keep only time series data for the id over which we are iterating
        ts_data_one_location = ts_data.loc[ts_data["pickup_location_id"] == location_id,
                                           ["pickup_hour", "rides"]
                                           ]
        
        # Pre-compute cutoff indices to split DataFrame rows
        indices = get_cutoff_indices(
            ts_data_one_location,
            input_seq_len,
            step_size
        )

        # Slice and transpose data into NumPy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, input_seq_len), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []
        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]["rides"].values
            y[i] = ts_data_one_location.iloc[idx[1] : idx[2]]["rides"].values[0]
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]["pickup_hour"])

        # Convert NumPy array to Pandas DataFrame
        features_one_location = pd.DataFrame(
            x,
            columns=[f"rides_previous_{i+1}_hour" for i in reversed(range(input_seq_len))]
            )
        features_one_location["pickup_hour"] = pickup_hours
        features_one_location["pickup_location_id"] = location_id

        # Convert NumPy target array to Pandas DataFrame
        targets_one_location = pd.DataFrame(y, columns=[f"target_rides_next_hour"])

        # Concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets["target_rides_next_hour"]

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*7*1, # Return one week of historical data
    step_size=24
)

print(f"{features.shape=}")
print(f"{targets.shape=}")

Our function returned the expected result. We have transformed our time-series into features and targets. We can move on to our next step.
