==> notebooks/04-transform_raw_data_into_features_and_targets.ipynb <==
# 04 | Transform Time Series into Features and Targets #

In this notebook, we will put together pieces of code that we implemented in the previous notebooks. We won't copy and paste the previous codes. For maintenance purposes, we will extract the functions from Python packages that we will create in our `src` directory.

# Always reload our imports to make sure they reflect changes.
%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data

We can now download available data for 2022 using our function.

rides = load_raw_data(year=2022)
rides

## Transform data into time series data ##
We can also import our premade functions. We'll use `add_missing_slots` and `transform_ts_data_into_ts_data`

from src.data import  transform_raw_data_into_ts_data

### Transform into time series data ###

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

We can now save our output in a new tabular data. We create a new `tabular_data` with our `features`. Then, we add a new column, `target_rides_next_hour`, which shall consist of our `targets`. Following, we can import our paths and save our outpu in the appropriate directory.

tabular_data  = features
tabular_data["target_rides_next_hour"] = targets

from src.paths import TRANSFORMED_DATA_DIR

tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")

After this step, we are done with the data transformation stage. We can move on to the following stage, which is building models (yay!)!
