==> notebooks/05-visualize_training_data.ipynb <==
# 04 | Visualize Training Data #

In this notebook, we will make a brief visual exploration of our training data to get a gist of it.

## Import features and target data ##

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
from src.plot import plot_one_sample

features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")

features_and_target

We can split the data into two different variables, `features` and `targets`.

features  = features_and_target.drop(columns=["target_rides_next_hour"])
targets = features_and_target["target_rides_next_hour"]

features_and_target.loc[features_and_target["pickup_location_id"] == 4].head(20)

## Visualize the generated training data ##

plot_one_sample(
    example_id=1011,
    features=features,
    targets=targets
)

Great, we are ready to develop our ML models now!
