==> notebooks/07-xboost_model.ipynb <==
# 07 | Xboost model #
## Introduction ##
In this notebook, we will try to improve our baseline model by implementing **machine learning models**. When it comes to tabular data and real world datasets, in most cases, the best algorithms are **boosting trees**.
- `XGBoost`
- `LightGBM`
- `CatBoost`

In this notebook, we will train the first algorithm - `XGBoost`.
## Load the training data ##

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
from datetime import datetime
from src.data_split import train_test_split
import xgboost as xgb
from sklearn.metrics import mean_absolute_error

## Load the data ##

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

## Split the data into train and test sets ##

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0), # Cutoff date = 06 june, 2022
    target_column_name="target_rides_next_hour"
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

We now only use past rides data as our training set. Consider that in our dataset, all previous rides start with `"rides_"`, but if we pass `pickup_hour` too, we will get the date column as well. For the moment, we will skip this column. 

# Use only past rides data
past_rides_columns = [c for c in X_train.columns if c.startswith("rides_")]
X_train_only_numeric = X_train[past_rides_columns]

Now, we will train the model. We instantiate an `XGBRegressor()` class with our features and target variables.

model = xgb.XGBRegressor()
model.fit(X_train_only_numeric, y_train)

We can now evaluate the model. First, though, we need to generate predictions using the test data by using the same columns as train data.

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

Let's compute the mean absolute error now.

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

Observations:
- The error is `2.70`. It is already lower than our baseline model. Remember? It was  **3.1963**.
- The algorithm is bringing complexity but it is bringing improvement.
- The problem: our model is kinda **heavy**. We have too many features, `674`.

We can try a different approach in the next feature by using a lighter, more efficient solution.
