==> notebooks/09-lightgbm_model_with_feature_engineering.ipynb <==
# 09 | LightGBM Model with Feature Engineering #
## Introduction ##
In this notebook, we will try to improve our baseline model by implementing **machine learning models**. When it comes to tabular data and real world datasets, in most cases, the best algorithms are **boosting trees**.
- `XGBoost`
- `LightGBM`
- `CatBoost`

In this notebook, we will train the **second** algorithm - `LightGBM` - again. This time, though, we will apply **feature engineering**. Feature engineering is about **expanding** the information that your input features have. This means adding columns to the dataset.
Ideas
1. Here, we will create a `avg_rides_last_4_weeks` feature, which is based on our third baseline model. It returened the average number of rides observed 7, 14, 21, and 28 days ago.
2. Derive numerical features from `pickup_hour`. We can create `hour` and `week_day` as new columns.
3. Use `pickup_location_id` (challenge). As the values are categorical, we can create some sort of categorical encoding. Instead of using those numbers, we could replace them by the `latitude` and `longitude` of some middle point in the area. We could provide a representation of the data that respects distances on a map and could be useful for our model.

## Load the training data ##

# Always reload our imports to make sure they reflect changes.
%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
from datetime import datetime
from src.data_split import train_test_split
import lightgbm as lgb
from sklearn.metrics import mean_absolute_error
from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.pipeline import make_pipeline

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

## Split the data into train and test sets ##

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name="target_rides_next_hour"
)

print(f"{X_train.shape=}")
print(f"{y_train.shape=}")
print(f"{X_test.shape=}")
print(f"{y_test.shape=}")

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from:

    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago

    Args:

    - `X` (`pd.DataFrame`): Pandas DataFrame with features

    Returns:

    - `pd.DataFrame` with extra column `average_rides_last_4_weeks`
    """
    X["average_rides_last_4_weeks"] = 0.25 * (
        X[f"rides_previous_{7*24}_hour"] + \
        X[f"rides_previous_{2*7*24}_hour"] + \
        X[f"rides_previous_{3*7*24}_hour"] + \
        X[f"rides_previous_{4*7*24}_hour"]   
    )

    return X

Becasue `average_rides_last_4_weeks` is a Python function, we need to transform it into a `scikit-learn` data transformation step using `FunctionTransformer`.

add_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False
)

add_average_rides_last_4_weeks.fit_transform(X_train)

We can now extract the hour of the day and the day of the week. We will use another way to create a data transformation step in `scikit-learn`. We can create a class `TemporalFeaturesEngineer` that inherits `BaseEstimator` and `TransformerMixin` from `scikit-learn`. We then implement two methods, `fit()` and `transform()`.
- If the data transformation has parameters, `fit()` needs to be implemented (not our case).
- `transform()` needs to pass the set of features X.

Remember: `LGBM` only accepts numerical values!

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):

        X_ = X.copy()

        # Generate numeric columns from datetime
        X_["hour"] = X_["pickup_hour"].dt.hour
        X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek

        return X_.drop(columns=["pickup_hour"])

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

We can now implement `make_pipeline` from `scikit-learn`, which will create a pipeline of events that on which we can implement transformations and predictions.

pipeline = make_pipeline(
    add_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)


test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

Observtions:
- After some feature engineering, our model did not improve. It actually got a bit worse, with a `0.011` difference, roughly. 
- Feature engineering is an experimental process.

Next, we will try to tackle the challenge that the instructor proposed to observe the difference in results.
